+++ seeders/DatabaseSeeder.ts
@@ -28,7 +28,7 @@ import { User } from '@/users/entities/user.entity';
 import { getDefaultModel } from '@/runs/execution/factory';
 import { SystemTools } from '@/tools/entities/tool-calls/system-call.entity';
 import { ProjectApiKey } from '@/administration/entities/project-api-key.entity';
-import { API_KEY_PREFIX, scryptApiKey } from '@/auth/utils';
+import { API_KEY_PREFIX, scryptSecret } from '@/auth/utils';
 import {
   ORGANIZATION_ID_DEFAULT,
   ORGANIZATION_OWNER_ID_DEFAULT,
@@ -36,6 +36,7 @@ import {
   PROJECT_ID_DEFAULT
 } from '@/config';
 import { redactProjectKeyValue } from '@/administration/helpers';
+import { Agent } from '@/runs/execution/constants';
 
 const USER_EXTERNAL_ID = 'test';
 const PROJECT_API_KEY = `${API_KEY_PREFIX}testkey`;
@@ -83,14 +84,15 @@ export class DatabaseSeeder extends Seeder {
       .getRepository(ProjectPrincipal)
       .getReference(projectUser.id, { wrapped: true });
     const projectApiKey = new ProjectApiKey({
-      key: scryptApiKey(PROJECT_API_KEY),
+      key: scryptSecret(PROJECT_API_KEY),
       name: 'test key',
       createdBy: ref(projectUser),
       project: ref(project),
       redactedValue: redactProjectKeyValue(PROJECT_API_KEY)
     });
-    const assistant = new Assistant({
+    const beeAssistant = new Assistant({
       model: getDefaultModel(),
+      agent: Agent.BEE,
       tools: [
         {
           type: 'system',
@@ -117,7 +119,20 @@ export class DatabaseSeeder extends Seeder {
         $ui_icon: 'Bee'
       }
     });
-    em.persist([assistant, projectApiKey]);
+    const streamlitAssistant = new Assistant({
+      model: getDefaultModel(),
+      agent: Agent.STREAMLIT,
+      tools: [],
+      name: 'Builder Assistant',
+      project: ref(project),
+      createdBy: ref(projectUser),
+      description: 'An example streamlit agent, tailored for building Streamlit applications.',
+      metadata: {
+        $ui_color: 'white',
+        $ui_icon: 'Bee'
+      }
+    });
+    em.persist([beeAssistant, streamlitAssistant, projectApiKey]);
     await em.flush();
     process.env.IN_SEEDER = undefined;
   }