+++ src/runs/execution/execute.ts
@@ -15,14 +15,14 @@
  */
 
 import { Loaded } from '@mikro-orm/core';
-import { TokenMemory } from 'bee-agent-framework/memory/tokenMemory';
 import { BaseMessage } from 'bee-agent-framework/llms/primitives/message';
 import { Version } from 'bee-agent-framework';
-import { BeeAgent } from 'bee-agent-framework/agents/bee/agent';
 import { Summary } from 'prom-client';
 import dayjs from 'dayjs';
 import { isTruthy } from 'remeda';
 import { createObserveConnector } from 'bee-observe-connector';
+import { BeeAgent } from 'bee-agent-framework/agents/bee/agent';
+import { TokenMemory } from 'bee-agent-framework/memory/tokenMemory';
 
 import { Run } from '../entities/run.entity.js';
 import { toRunDto } from '../runs.service.js';
@@ -30,11 +30,10 @@ import { toRunDto } from '../runs.service.js';
 import {
   addFileToToolResource,
   checkFileExistsOnToolResource,
-  createToolResource,
-  getPromptTemplate
+  createToolResource
 } from './helpers.js';
 import { getTools } from './tools/helpers.js';
-import { createChatLLM } from './factory.js';
+import { createAgentRun, createChatLLM } from './factory.js';
 
 import { ORM } from '@/database.js';
 import { getLogger } from '@/logger.js';
@@ -47,7 +46,6 @@ import { Trace } from '@/observe/entities/trace.entity.js';
 import { RunStep } from '@/run-steps/entities/run-step.entity.js';
 import { Message } from '@/messages/message.entity.js';
 import { AnyToolCall } from '@/tools/entities/tool-calls/tool-call.entity.js';
-import { createStreamingHandler } from '@/runs/execution/event-handlers/streaming.js';
 import { LoadedRun } from '@/runs/execution/types.js';
 import { UserResource } from '@/tools/entities/tool-resources/user-resource.entity.js';
 import { SystemResource } from '@/tools/entities/tool-resources/system-resource.entity.js';
@@ -62,7 +60,7 @@ const agentExecutionTime = new Summary({
 });
 
 export type AgentContext = {
-  run: Loaded<Run>;
+  run: Loaded<Run, 'assistant'>;
   publish: ReturnType<typeof createPublisher>;
   runStep?: Loaded<RunStep>;
   message?: Loaded<Message>;
@@ -71,8 +69,6 @@ export type AgentContext = {
 
 export async function executeRun(run: LoadedRun) {
   const runLogger = getLogger().child({ runId: run.id });
-
-  const llm = createChatLLM(run);
 
   const publish = createPublisher(run);
 
@@ -125,25 +121,15 @@ export async function executeRun(run: LoadedRun) {
     run.thread.$.toolResources = toolResources;
   }
 
-  const memory = new TokenMemory({
-    llm
-  });
-  await memory.addMany(messages);
-
   run.start();
   await ORM.em.flush();
   await publish({ event: 'thread.run.in_progress', data: toRunDto(run) });
-
   const context = { run, publish } as AgentContext;
 
-  const agent = new BeeAgent({
-    llm,
-    memory,
-    tools: await getTools(run, context),
-    templates: {
-      system: getPromptTemplate(run)
-    }
-  });
+  const tools = await getTools(run, context);
+  const llm = createChatLLM(run);
+  const memory = new TokenMemory({ llm });
+  await memory.addMany(messages);
 
   const cancellationController = new AbortController();
   const unsub = watchForCancellation(Run, run, () => cancellationController.abort());
@@ -152,25 +138,18 @@ export async function executeRun(run: LoadedRun) {
 
   try {
     const endAgentExecutionTimer = agentExecutionTime.labels({ framework: Version }).startTimer();
-    const agentRunPromise = agent
-      .run(
-        { prompt: null }, // messages have been loaded to agent's memory
-        {
-          // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-          // @ts-ignore
-          signal: AbortSignal.any([cancellationController.signal, expirationSignal]),
-          execution: {
-            totalMaxRetries: 10,
-            maxRetriesPerStep: 3,
-            maxIterations: 10
-          }
-        }
-      )
-      .observe(createStreamingHandler(context));
+    const [agentRun, agent] = createAgentRun(
+      run,
+      { llm, tools, memory },
+      {
+        signal: AbortSignal.any([cancellationController.signal, expirationSignal]),
+        ctx: context
+      }
+    );
 
     // apply observe middleware only when the observe API is enabled
-    if (BEE_OBSERVE_API_URL && BEE_OBSERVE_API_AUTH_KEY) {
-      agentRunPromise.middleware(
+    if (BEE_OBSERVE_API_URL && BEE_OBSERVE_API_AUTH_KEY && agent instanceof BeeAgent) {
+      (agentRun as ReturnType<BeeAgent['run']>).middleware(
         createObserveConnector({
           api: {
             baseUrl: BEE_OBSERVE_API_URL,
@@ -196,7 +175,7 @@ export async function executeRun(run: LoadedRun) {
       );
     }
 
-    await agentRunPromise;
+    await agentRun;
 
     endAgentExecutionTimer();
     run.complete();
@@ -224,6 +203,5 @@ export async function executeRun(run: LoadedRun) {
   } finally {
     await publish({ event: 'done', data: '[DONE]' });
     unsub();
-    agent.destroy();
   }
 }