+++ src/runs/execution/helpers.ts
@@ -15,10 +15,7 @@
  */
 
 import { Loaded, ref, Ref } from '@mikro-orm/core';
-import { BeeSystemPrompt } from 'bee-agent-framework/agents/bee/prompts';
 import { unique } from 'remeda';
-
-import { Run } from '../entities/run.entity.js';
 
 import { ORM } from '@/database.js';
 import { File } from '@/files/entities/file.entity.js';
@@ -45,22 +42,6 @@ export function getRunVectorStores(
     )
     .flatMap((container) => container.vectorStore.$);
   return unique(vectorStores); // filter out duplicates
-}
-
-export function getPromptTemplate(run: Loaded<Run, 'assistant'>) {
-  const instructions = run.additionalInstructions
-    ? `${run.instructions} ${run.additionalInstructions}`
-    : run.instructions;
-  return BeeSystemPrompt.fork((input) => ({
-    ...input,
-    ...(run.assistant.$.systemPromptOverwrite
-      ? { template: run.assistant.$.systemPromptOverwrite }
-      : {}),
-    defaults: {
-      ...input.defaults,
-      ...(instructions ? { instructions } : {})
-    }
-  }));
 }
 
 export async function checkFileExistsOnToolResource(